+++ MaxText/tests/attention_test.py
@@ -39,7 +39,8 @@ from MaxText import pyconfig
 from MaxText.common_types import DECODING_ACTIVE_SEQUENCE_INDICATOR, MODEL_MODE_AUTOREGRESSIVE, MODEL_MODE_PREFILL, MODEL_MODE_TRAIN
 from MaxText.globals import PKG_DIR
 from MaxText.layers import attentions
-from MaxText.layers.attentions import Attention, MLA, ChunkedCausalMask
+from MaxText.layers.attentions import Attention, ChunkedCausalMask
+from MaxText.layers.attention_mla import MLA
 
 
 class BidirectionalBlockMaskTest(unittest.TestCase):
